# api-specification-singulary.yaml
openapi: 3.0.3
info:
  title: Singulary API
  description: |
    API completa para o aplicativo Singulary - Ressonador Quântico Pessoal (RQP).
    
    Esta API fornece endpoints para gerenciamento de usuários, presets de frequências,
    sessões de meditação/terapia sonora e integração futura com hardware.
    
    ## Funcionalidades Principais
    - Autenticação JWT
    - Gerenciamento de presets (padrão e personalizados)
    - Registro e análise de sessões
    - Exportação de dados em CSV
    - Estatísticas e relatórios
    - Preparação para integração com hardware
    
    ## Segurança
    - Autenticação via JWT tokens
    - Validação de dados de entrada
    - Rate limiting (implementação futura)
    - CORS configurado para frontend
    
    ## Versioning
    Esta é a versão 1.0 da API, compatível com o MVP do Singulary.
    
  version: 1.0.0
  contact:
    name: Equipe Singulary
    email: api@singulary.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desenvolvimento local
  - url: https://api.singulary.app/v1
    description: Servidor de produção (futuro)
...
paths:
...
  /presets:
    get:
      tags:
        - presets
      summary: Listar presets
      description: |
        Obter lista de presets disponíveis (padrão + personalizados do usuário).
        Inclui filtros opcionais por dimensão e categoria.
      parameters:
        - name: user_id
          in: query
          description: ID do usuário para incluir presets personalizados
          schema:
            type: string
            example: user-abc123
        - name: dimension
          in: query
          description: Filtrar por dimensão
          schema:
            type: string
            enum: [3D, 4D, 5D, Personalizado]
            example: 3D
        - name: category
          in: query
          description: Filtrar por categoria
          schema:
            type: string
            enum: [Foco, Relaxamento, Meditação, Sono, Espiritual, Cura, Personalizado]
            example: Foco
      responses:
        '200':
          description: Lista de presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Preset'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - presets
      summary: Criar preset personalizado
      description: Criar um novo preset personalizado para o usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - baseFrequency
                - beatFrequency
                - lightFrequency
                - duration
              properties:
                name:
                  type: string
                  example: Minha Frequência Especial
                  description: Nome do preset
                description:
                  type: string
                  example: Preset personalizado para concentração profunda
                  description: Descrição opcional
                dimension:
                  type: string
                  default: Personalizado
                  example: Personalizado
                category:
                  type: string
                  default: Personalizado
                  example: Foco
                baseFrequency:
                  type: number
                  minimum: 100
                  maximum: 1000
                  example: 528
                  description: Frequência base em Hz
                beatFrequency:
                  type: number
                  minimum: 1
                  maximum: 100
                  example: 40
                  description: Frequência de batida binaural em Hz
                lightFrequency:
                  type: number
                  minimum: 1
                  maximum: 30
                  example: 10
                  description: Frequência de luz estroboscópica em Hz
                duration:
                  type: integer
                  minimum: 1
                  maximum: 120
                  example: 20
                  description: Duração padrão em minutos
                userId:
                  type: string
                  example: user-abc123
                  description: ID do usuário (opcional, pode ser extraído do token)
      responses:
        '201':
          description: Preset criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /presets/{presetId}:
    get:
      tags:
        - presets
      summary: Obter preset específico
      description: Obter detalhes de um preset específico
      parameters:
        - name: presetId
          in: path
          required: true
          description: ID do preset
          schema:
            type: string
            example: 3d-focus
      responses:
        '200':
          description: Detalhes do preset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preset'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - presets
      summary: Atualizar preset personalizado
      description: Atualizar um preset personalizado (apenas presets criados pelo usuário)
      security:
        - bearerAuth: []
      parameters:
        - name: presetId
          in: path
          required: true
          description: ID do preset
          schema:
            type: string
            example: custom-abc123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Minha Frequência Atualizada
                description:
                  type: string
                  example: Descrição atualizada
                baseFrequency:
                  type: number
                  minimum: 100
                  maximum: 1000
                  example: 432
                beatFrequency:
                  type: number
                  minimum: 1
                  maximum: 100
                  example: 8
                lightFrequency:
                  type: number
                  minimum: 1
                  maximum: 30
                  example: 4
                duration:
                  type: integer
                  minimum: 1
                  maximum: 120
                  example: 30
                category:
                  type: string
                  example: Relaxamento
      responses:
        '200':
          description: Preset atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Não é possível editar presets padrão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - presets
      summary: Deletar preset personalizado
      description: Deletar um preset personalizado (apenas presets criados pelo usuário)
      security:
        - bearerAuth: []
      parameters:
        - name: presetId
          in: path
          required: true
          description: ID do preset
          schema:
            type: string
            example: custom-abc123
      responses:
        '200':
          description: Preset deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Preset deletado com sucesso
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Não é possível deletar presets padrão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /presets/seed:
    post:
      tags:
        - presets
      summary: Criar presets padrão
      description: |
        Endpoint para criar presets padrão do sistema (apenas para desenvolvimento).
        Cria os presets básicos das dimensões 3D, 4D e 5D.
      responses:
        '200':
          description: Presets padrão criados ou já existem
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 6 presets padrão criados com sucesso
        '500':
          $ref: '#/components/responses/InternalServerError'

  # SESSÕES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido através do endpoint de login.
        Formato: `Bearer <token>`
  schemas:
...
    Preset:
      type: object
      properties:
        id:
          type: string
          example: 3d-focus
          description: ID único do preset
        name:
          type: string
          example: Foco 3D
          description: Nome do preset
        description:
          type: string
          example: Frequências para concentração e produtividade
          description: Descrição do preset
        dimension:
          type: string
          enum: [3D, 4D, 5D, Personalizado]
          example: 3D
          description: Dimensão espiritual do preset
        category:
          type: string
          enum: [Foco, Relaxamento, Meditação, Sono, Espiritual, Cura, Personalizado]
          example: Foco
          description: Categoria do preset
        baseFrequency:
          type: number
          example: 528
          description: Frequência base em Hz
        beatFrequency:
          type: number
          example: 40
          description: Frequência de batida binaural em Hz
        lightFrequency:
          type: number
          example: 10
          description: Frequência de luz estroboscópica em Hz
        duration:
          type: integer
          example: 20
          description: Duração padrão em minutos
        isCustom:
          type: boolean
          example: false
          description: Se é um preset personalizado
        userId:
          type: string
          nullable: true
          example: user-abc123
          description: ID do usuário (null para presets padrão)
        createdAt:
          type: string
          format: date-time
          example: 2025-08-01T10:00:00Z
          description: Data de criação
        updatedAt:
          type: string
          format: date-time
          example: 2025-08-10T15:30:00Z
          description: Data da última atualização
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
          description: Descrição do erro
        code:
          type: string
          example: INVALID_INPUT
          description: Código do erro (opcional)
        details:
          type: object
          description: Detalhes adicionais do erro (opcional)
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Dados de entrada inválidos
            code: INVALID_INPUT
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Token inválido ou expirado
            code: UNAUTHORIZED
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso não encontrado
            code: NOT_FOUND
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Erro interno do servidor
            code: INTERNAL_ERROR
...